Updated configuration files.
Updated log aliases for Git for a more informative output. Set the C++ checker
for Vim's Syntastic plugin so that it wouldn't default to the g++ compiler.
Lastly updated the bash alias for connecting to my VPN so that it's in
the correct directory when using the VPN's configuration file (Which as
relative file paths.)