Added missing options to Git and Vim configurations.
Added the show signature option to my Git's one-line formatted output so
that it's easier to see which commits are properly signed.

Also added the node_modules, bower_components, and dist directories to
the list of directories for the CtrlP plugin to ignore when building
its cache (As these directories are either build artifacts or external
libraries).