Updates to Bash, Git, and Vim configuration files.
Updated Bash configuration to properly set the TERM environmental
variable to indicate that color is supported by the terminal. That is
required by terminal-based applications such as Vim and TMUX when those
applications want to use themes, or color schemes, that require
knowledge of how many colors are supported by the terminal

Updated Git's configuration file so that commits are automatically
cryptographically signed.

Updated Vim's configuration file to actually use a custom color scheme
(As that feature wasn't working properly earlier as a result of
conflicting Vim setup steps.). Also added a new plugin that makes
hundreds of color schemes available for selection through the
:colorscheme command in Vim. Lastly added, and configured, vim-signify,
which brings with it support for +, -, and ~ characters in the left
sidebar indicating changes to a file under version control.