Adding new Vim plugins and setting git option.
Replaced the vim-colorscheme plugin with the jellybeans plugin. This was necessary to resolve an issue with the vim-colorscheme plugin. The vim-colorscheme plugin has an old version of the jellybeans colorscheme that fails to properly work with vimdiff (The older version does not highlight changes in red or green).

Also added a new plugin called tmuxline to setup and configure a TMUX status line using the colorscheme of the current colorscheme being used by Vim. Effectively this plugin manages the status line of TMUX. It only works, however, if currently in a TMUX window, and only after starting Vim at least once.

Lastly I added another option to my Git configuration file to use mnemonic branch names when generating Git diffs rather than the standard 'a' and 'b' branch names.