Alter behavior of CtrlP.
Limit recursion by CtrlP from a directory depth of 40 to a directory
depth of 20.

Add 'build' directories to the list of directories that should be
ignored by CtrlP. This will account for the use of directories named
'build' used by the build processes of C++ projects. We don't
care to load build artifacts into Vim.

Restrict CtrlP to only match against files, and exclude buffers and MRU
files. This will help restrict CtrlP to the directory of the working
project, where it's most likely a user is interested in pulling up a
files.