Enable markdown support for Tagbar.
Add third-party repository as a submodule of this repository. This
third-party repository contains a script capable of generating
Ctag information required by Tagbar to display heading information.

Update deployment script to symlink the third-party script into the
user's Vim directory so that it can be referenced later.

Add additional configuration information to user's `.vimrc` to instruct
Tagbar to call the third-party script to get the necessary Ctag
information required to display a tree view of a Markdown file's
headings.