Set additional merge options for Git.
Set the `conflictstyle` option so that Git will updating the working
tree file with the contents of the two sides of a merge and the original
file.

Set the `renormalize` option to force normalization of files on either
side of a merge so that changes in line endings, or other attributes, do
not cause unnecessary conflicts.